[refactor]: navigation 공통 component를 재활용하도록 수정(개수추가는 선택적으로)
